+++ ecms_base/features/custom/ecms_speeches/tests/src/Functional/EcmsSpeechesInstallTest.php
@@ -30,7 +30,7 @@ class EcmsSpeechesInstallTest extends AllProfileInstallationTestsAbstract {
    *
    * @var string
    */
-  protected $defaultTheme = 'classy';
+  protected $defaultTheme = 'ecms';
 
   /**
    * Test the ecms_speeches installation.