+++ CHANGELOG.md
@@ -45,6 +45,7 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 - RIG-45: Added Executive Order content type as optional feature.
 - RIG-72: Added the queue and publish service to syndicate content to other sites.
 - RIG-67: Added paragraphs feature to base install.
+- RIG-62: Added webform node module install by default.
 
 ### Changed
 - RIG-23: Changed from OIDC generic to Windows AAD for authentication.